refactor: remove unused `ProjectManager.manifest` property
This property is unused and when I was experimenting in the console its
existence confused me. It's using `self.sources` to build the
`PackageManifest`, which is `List[Path]` when it should be `Dict[str,
Source]`. I also think it's confusing to have a `manifest` and
`cached_manifest` property. I'll add this property back if its needed.