fix: allow Gnosis Safe style proxies to work with earlier deployments
Prior to v1.1.0, `masterCopy` was not available as an external method.
This means that our proxy detection would not be able to detect it as
a Gnosis Safe-style proxy. Additionally, once deployed, this style of
proxy is immutable, so even if the Safe itself is upgraded to a newer
version, this method still would not be available to be called.

NOTE: This slightly modifies the proxy info's target to be obtained
      dynamically from retreiving the storage for `masterCopy` directly.